Ajustes para la implementacion de casos de prueba
