+++ helpers/verifyJWT.js
@@ -3,12 +3,14 @@ import jwt from 'jsonwebtoken';
 const verifyJWT = (request, response) => {
     const token = request.headers['authorization'];
     if (!token) {
-        return response.status(403).send('Acceso denegado');
+        response.status(403).send('Acceso denegado');
+        return;
     }
 
     jwt.verify(token, process.env.JWT_SECRET, (error, data) => {
         if (error) {
-            return response.status(401).send('Acceso denegado');
+            response.status(401).send('Acceso denegado');
+            return;
         }
         request.userTKN = data;
     });
